PF-28 Moved the UriHelper class to a common Helpers namespace
