+++ src/Aquarius.Client/Aquarius.Client.csproj
@@ -65,7 +65,7 @@
     <Compile Include="TimeSeries\Client\EndPoints\Provisioning.cs" />
     <Compile Include="TimeSeries\Client\EndPoints\PublishV2.cs" />
     <Compile Include="TimeSeries\Client\EndPoints\Root.cs" />
-    <Compile Include="TimeSeries\Client\EndPoints\UriHelper.cs" />
+    <Compile Include="Helpers\UriHelper.cs" />
     <Compile Include="TimeSeries\Client\Helpers\AuthenticationHeaders.cs" />
     <Compile Include="TimeSeries\Client\Helpers\ClientHelper.cs" />
     <Compile Include="TimeSeries\Client\Helpers\DurationExtensions.cs" />
