+++ src/Aquarius.Client/TimeSeries/Client/AquariusSystemDetector.cs
@@ -4,6 +4,7 @@ using System.Configuration;
 using System.Diagnostics;
 using System.Net;
 using System.Reflection;
+using Aquarius.Helpers;
 using Aquarius.TimeSeries.Client.EndPoints;
 using ServiceStack;
 using ServiceStack.Logging;
