+++ src/Aquarius.Client/TimeSeries/Client/EndPoints/PublishV2.cs
@@ -1,4 +1,6 @@
-﻿namespace Aquarius.TimeSeries.Client.EndPoints
+﻿using Aquarius.Helpers;
+
+namespace Aquarius.TimeSeries.Client.EndPoints
 {
     public class PublishV2
     {
